two small warning fixes
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@484 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/projectM-engine/PCM.hpp
@@ -36,7 +36,7 @@ public:
     int start;
 
     /** Use wave smoothing */
-    int waveSmoothing;
+    float waveSmoothing;
 
     int *ip;
     double *w;
